Move PurchaseExpiringNotice to own file
